refactor: simplify frontend module and ownership layout
